Ignore python_requires when installing with distutils
When using python setup.py <command>, distutils.dist.Distribution
does not recognize python_requires (unlike setuptools) and will
issue an obnoxious warning that doesn't actually tell the user
that their Python version is not supported.

Note that this (might not be) the case with pip since it replaces
distutils with setuptools, which does support it.

The other possible solution would be to parse python_requires,
but that requires a lot of work (see pip/_internals/utils/packaging.py)
and is an example of the reason setuptools was made in the first place.