Allow replacement of Python Loader/Dumper with C versions
Some pyyaml users prefer the fast performance of libyaml, and are less
concerned about minor incompatibilities. There is no existing method to
globally enable the libyaml bindings, and different libraries may do
things in different ways.

This change provides a way for users to request global replacement of
the default Python Loaders and Dumpers with the libyaml bindings, prior
to the initial import of the pyyaml library.

To enable in Python3:
    import builtins
    builtins._yaml__use_libyaml_by_default__ = True
    import yaml

To enable in Python2:
    import __builtin__
    __builtin__._yaml__use_libyaml_by_default__ = True
    import yaml

Assuming that libyaml is installed, the result will be that:
    yaml.BaseLoader == yaml.cyaml.CBaseLoader
    yaml.SafeLoader == yaml.cyaml.CSafeLoader
    yaml.FullLoader == yaml.cyaml.CFullLoader
    yaml.UnsafeLoader == yaml.cyaml.CUnsafeLoader
    yaml.Loader == yaml.cyaml.CLoader
    yaml.BaseDumper == yaml.cyaml.CBaseDumper
    yaml.SafeDumper == yaml.cyaml.CSafeDumper
    yaml.Dumper == yaml.cyaml.CDumper

This allows the use of the "sugar" methods (safe_load, full_load,
unsafe_load, etc.) as is, benefiting from the performance boost of
libyaml if it is installed. This should also provide very broad
compatibility with existing code that may explicitly reference specific
loaders and dumpers by name.