Initial drop of libfyaml based backend.
This module implements an additional C based backend based on
libfyaml.

For the initial drop the goal is to have something as similar as
possible to the libyaml backend, where this is possible.

It is still in an experimental phase, so things like error
support do not work yet.

In order to use replace the capital C with an F on each case,
i.e. instead of CLoader use FLoader, CDumper use FDumper etc.

Signed-off-by: Pantelis Antoniou <pantelis.antoniou@konsulko.com>