Use `ChainMap` to improve representer/constructor inheritance
Before this change, a child Dumper "inherited" its parent Dumper's
representers by simply copying the parent's representers the first
time a new representer was registered with the child Dumper.  This
approach works as users expect only if representers are never added to
a parent (ancestor) Dumper after representers are added to a
child (descendant) Dumper.  Same goes for Loaders and their registered
constructors.

This commit uses a `collections.ChainMap` built from ancestor `dict`
objects in method resolution order (MRO) to provide true inheritance
of representers (for Dumpers) and constructors (for Loaders).

This is technically a backwards-incompatible change: This change
breaks any code that intentionally subverts the expected inheritance
behavior by registering a function in an ancestor class *after*
registering a function in a descendant class.